fixed invalid date on the DatePicker's CalendarHeader this is a small fix to prevent "Invalid date" from showing in the CalendarHeader
this happens when the `date` parameter is a moment object that is invalid.

FWIW in separate implementations, we found that the `decrementYear` / `decrementMonth` / `incrementMonth` / `incrementYear` functions would try to modify a `null` value like it was a moment object, rendering it invalid.

if an invalid date is passed in, it breaks the variables responding for printing the month & year in the calendar picker

this is fixed by checking inside the component if the `date` property is invalid, and if so, reverting to the current date / year

this is not ideal, as if someone tries to increment/decrement a month or year that represents now, that indicates they are choosing the date, with that in mind, I think it's worth setting the value on the date as if they were transforming from the default value, but this cannot be done without moving the decrement & increment functions inside the function, as they have no variables in their signature from material-ui or onwards

this is a breaking API change for the DatePicker, so it makes sense to fix separately (and bump minor version)